refactor: Role enum 리팩토링
