+++ src/main/java/org/indoles/memberserviceserver/core/domain/enums/Role.java
@@ -1,0 +1,27 @@
+package org.indoles.memberserviceserver.core.domain.enums;
+
+import org.indoles.memberserviceserver.global.exception.BadRequestException;
+
+import java.util.Arrays;
+
+import static org.indoles.memberserviceserver.global.exception.ErrorCode.M001;
+
+public enum Role {
+
+    SELLER("판매자"),
+    BUYER("입찰자(구매자");
+
+    private final String description;
+
+    Role(final String description) {
+        this.description = description;
+    }
+
+    public static Role find(final String userRole) {
+        return Arrays.stream(values())
+                .filter(role -> role.name().equals(userRole))
+                .findAny()
+                .orElseThrow(() -> new BadRequestException("사용자의 역할을 찾을 수 없습니다. userRole = " + userRole, M001));
+    }
+}
+
